+++ muk_web_searchpanel/__manifest__.py
@@ -16,38 +16,36 @@
 #    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #
 ###################################################################################
- 
-{ 
-    "name": "MuK Web Utils",
-    "summary": """Utility Features""",
-    "version": "12.0.2.8.22", 
-    "category": "Extra Tools",
-    "license": "AGPL-3",
-    "author": "MuK IT",
-    "website": "http://www.mukit.at",
+
+{
+    'name': 'MuK Search Panel',
+    'summary': 'Kanban Search Panel',
+    'version': '12.0.1.0.2',
+    'category': 'Extra Tools',
+    'license': 'AGPL-3',
+    'author': 'MuK IT',
+    'website': 'https://www.mukit.at',
     'live_test_url': 'https://mukit.at/r/SgN',
-    "contributors": [
-        "Mathias Markl <mathias.markl@mukit.at>",
+    'contributors': [
+        'Mathias Markl <mathias.markl@mukit.at>',
     ],
-    "depends": [
-        "web_editor",
-        "muk_utils",
+    'depends': [
+        'muk_web_utils',
     ],
-    "data": [
+    'data': [
         "template/assets.xml",
-        "views/res_config_settings_view.xml",
     ],
-    "qweb": [
-        "static/src/xml/*.xml",
+    'qweb': [
+        'static/src/xml/*.xml',
     ],
-    "images": [
+    'images': [
         'static/description/banner.png'
     ],
-    "external_dependencies": {
-        "python": [],
-        "bin": [],
+    'external_dependencies': {
+        'python': [],
+        'bin': [],
     },
-    "application": False,
-    "installable": True,
+    'application': False,
+    'installable': True,
     'auto_install': False,
-} 
+}