+++ muk_web_theme/models/res_config_settings.py
@@ -30,10 +30,26 @@ class ResConfigSettings(models.TransientModel):
 
     _inherit = 'res.config.settings'
 
+    #----------------------------------------------------------
+    # Database
+    #----------------------------------------------------------
+    
     theme_background_image = fields.Binary(
         related="company_id.background_image",
         readonly=False,
         required=True)
+    
+    theme_background_blend_mode = fields.Selection(
+        related="company_id.background_blend_mode",
+        readonly=False)
+    
+    theme_default_sidebar_preference = fields.Selection(
+        related="company_id.default_sidebar_preference",
+        readonly=False)
+
+    theme_default_chatter_preference = fields.Selection(
+        related="company_id.default_chatter_preference",
+        readonly=False)
     
     theme_color_brand = fields.Char(
         string="Theme Brand Color")
@@ -41,41 +57,63 @@ class ResConfigSettings(models.TransientModel):
     theme_color_primary = fields.Char(
         string="Theme Primary Color")
     
-    theme_color_appbar = fields.Char(
+    theme_color_menu = fields.Char(
+        string="Theme Menu Color")
+    
+    theme_color_appbar_color = fields.Char(
         string="Theme AppBar Color")
+    
+    theme_color_appbar_background = fields.Char(
+        string="Theme AppBar Background")
+    
+    #----------------------------------------------------------
+    # Functions
+    #----------------------------------------------------------
     
     @api.multi 
     def set_values(self):
         res = super(ResConfigSettings, self).set_values()
+        param = self.env['ir.config_parameter'].sudo()
         variables = [
             'o-brand-odoo',
             'o-brand-primary',
-            'mk-appbar-background'
+            'mk-apps-color',
+            'mk-appbar-color',
+            'mk-appbar-background',
         ]
         colors = self.env['muk_utils.scss_editor'].get_values(
             SCSS_URL, XML_ID, variables
         )
-        brand_changed = self.theme_color_brand != colors['o-brand-odoo']
-        primary_changed = self.theme_color_primary != colors['o-brand-primary']
-        appbar_changed = self.theme_color_appbar != colors['mk-appbar-background']
-        if(brand_changed or primary_changed or appbar_changed):
+        colors_changed = []
+        colors_changed.append(self.theme_color_brand != colors['o-brand-odoo'])
+        colors_changed.append(self.theme_color_primary != colors['o-brand-primary'])
+        colors_changed.append(self.theme_color_menu != colors['mk-apps-color'])
+        colors_changed.append(self.theme_color_appbar_color != colors['mk-appbar-color'])
+        colors_changed.append(self.theme_color_appbar_background != colors['mk-appbar-background'])
+        if(any(colors_changed)):
             variables = [
                 {'name': 'o-brand-odoo', 'value': self.theme_color_brand or "#706775"},
                 {'name': 'o-brand-primary', 'value': self.theme_color_primary or "#00A09D"},
                 {'name': 'mk-appbar-background', 'value': self.theme_color_appbar or "#706775"},
+                {'name': 'mk-apps-color', 'value': self.theme_color_menu or "#f8f9fa"},
+                {'name': 'mk-appbar-color', 'value': self.theme_color_appbar_color or "#dee2e6"},
             ]
             self.env['muk_utils.scss_editor'].replace_values(
                 SCSS_URL, XML_ID, variables
             )
+        param.set_param('muk_web_theme.background_blend_mode', self.theme_background_blend_mode)
         return res
 
     @api.model
     def get_values(self):
         res = super(ResConfigSettings, self).get_values()
+        params = self.env['ir.config_parameter'].sudo()
         variables = [
             'o-brand-odoo',
             'o-brand-primary',
-            'mk-appbar-background'
+            'mk-apps-color',
+            'mk-appbar-color',
+            'mk-appbar-background',
         ]
         colors = self.env['muk_utils.scss_editor'].get_values(
             SCSS_URL, XML_ID, variables
@@ -83,6 +121,9 @@ class ResConfigSettings(models.TransientModel):
         res.update({
             'theme_color_brand': colors['o-brand-odoo'],
             'theme_color_primary': colors['o-brand-primary'],
-            'theme_color_appbar': colors['mk-appbar-background'],
+            'theme_color_menu': colors['mk-apps-color'],
+            'theme_color_appbar_color': colors['mk-appbar-color'],
+            'theme_color_appbar_background': colors['mk-appbar-background'],
+            'theme_background_blend_mode': params.get_param('muk_web_theme.background_blend_mode', 'normal'),
         })
         return res